OnlineSVM pre-version.
Working on all the splits without segfaults and updating the weights. Also the weights of each iteration is not stored yet.
Java code is used to split the dataset. (Although the problem of segfaults was with not correctly passing the chunked datasets to find_max_violators_all and
infer_latent_all functions.. found from later diagnosis). But the Java code to split datasets will be handy in the future, to create randomised sub-sets etc. So
keeping this.

Only line 6 of the write-up is called multiple times. So this is a pre-version of the OnlineSVM. Still need to store the weights, appropriately modify the
objective function etc.